% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bund.R
\docType{data}
\name{bund}
\alias{bund}
\title{The first of three files comprising an example of a "Qualitaetsbericht".}
\format{
An XML-file with the following elements:
\describe{
  \item{Berichtsjahr}{E.g., "2019"}
  \item{IK_Krankenhaus}{E.g., "260100023"}
  \item{Standort}{E.g., "01"}
  \item{Land}{"DV"}
}
}
\source{
\url{https://www.g-ba.de/themen/qualitaetssicherung/datenerhebung-zur-qualitaetssicherung/datenerhebung-qualitaetsbericht/}
}
\usage{
bund
}
\description{
This file contains only marginal information about the year of the report, the
IK number ("Institutskennzeichen"), the number of the facility ("Standort")
and the variable "Land" ("Kürzel der Bundesländer der Geschäftsstellen").
}
\keyword{datasets}
