% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\docType{data}
\name{details}
\alias{details}
\title{The third of three files comprising an example of a "Qualitaetsbericht".}
\format{
An XML-file with the following elements (excerpt):
\describe{
  \item{Einleitung (with child elements)}{}
  \item{Krankenhaus (with child elements)}{}
  \item{Krankenhaustraeger (with child elements)}{}
  ...
}
}
\source{
\url{https://www.g-ba.de/themen/qualitaetssicherung/datenerhebung-zur-qualitaetssicherung/datenerhebung-qualitaetsbericht/}
}
\usage{
details
}
\description{
This file contains all the details that are included in a "Qualitaetsbericht",
that is the raw data about the clinic. This includes the contact data of the
persons responsible for the clinic and its departments, the amount of diagnoses
and procedures given or carried out, respectively, the medical and non-medical
services available and much more.
}
\keyword{datasets}
