% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\docType{data}
\name{land}
\alias{land}
\title{The second of three files comprising an example of a "Qualitaetsbericht".}
\format{
An XML-file with the following elements (excerpt):
\describe{
  \item{Berichtsjahr}{E.g., "2019"}
  \item{IK_Krankenhaus}{E.g., "260100023"}
  \item{Standort}{E.g., "01"}
  \item{Land}{E.g., "SH"}
  \item{Leistungsbereich (with child elements)}{}
  \item{QS-Ergebnis (with child elements)}{}
}
}
\source{
\url{https://www.g-ba.de/themen/qualitaetssicherung/datenerhebung-zur-qualitaetssicherung/datenerhebung-qualitaetsbericht/}
}
\usage{
land
}
\description{
This file contains the results of the "Qualitaetssicherung" regarding all
"Leistungsbereiche" that are offered in the clinic, e.g., for procedures
as "Herzschrittmacher-Implantation".
}
\keyword{datasets}
