% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_insert_single_table.R
\name{lists_insert_single_table}
\alias{lists_insert_single_table}
\title{Generate the correct table from selective lists to be inserted into the database.}
\usage{
lists_insert_single_table(
  table_name,
  primary_key_var,
  variable_lookup_table,
  search_env_for_lists = .GlobalEnv
)
}
\arguments{
\item{table_name}{A single string value with the name of the list.}

\item{primary_key_var}{A single string value that specifies the name of the
primary key of the database table.}

\item{variable_lookup_table}{A named list with single character values as
elements that specify which original variable names in the selective lists
(which are the list elements) map to which variable names in the database
(which are the names of the list elements).}

\item{search_env_for_lists}{The environment which is used to search for the
selective list objects, e.g., \code{lists_2015}. Defaults is the \code{.GlobalEnv}.}
}
\value{
A \code{tibble} with as many columns as the selective list has variables,
    and the primary key as specified in \code{primary_key_var} and the foreign
    key variable \code{HospitalDataYear_idHospitalDataYear}. The variables
    specified in \code{variable_lookup_table} are renamed accordingly.
}
\description{
Generate the correct table from selective lists to be inserted into the database.
}
\details{
The selective list objects as obtained via calls to
    \code{\link{qb_extract_selective_lists}} must be present, either in the
    global environment (the default value for the parameter
    \code{search_env_for_lists}) or in another environment,
    and must be named following the pattern \code{lists_2015} or
    \code{lists_2016} (\code{lists_20xx}).
}
\examples{
\dontrun{
path2015 <- "./00-data/Auswahllisten/2015_8-5-5_5_2016-06-21_Auswahllisten_BJ_2015.xls"
lists_2015 <- qb_extract_selective_lists(path2015)
path2016 <- "./00-data/Auswahllisten/2017-06-14_Anlage_Auswahllisten_Anhang-2_Anl-1_2016.xls"
lists_2016 <- qb_extract_selective_lists(path2016)

df <- lists_insert_single_table(table_name = "Medizinisch_pflegerische_Leistungsangebote",
                                primary_key_var = "idqb_list_nursingservices",
                                variable_lookup_table = list("code" = "Nummer",
                                   "description" = "Medizinisch_pflegerisches_Leistungsangebot",
                                   "comment" = "Kommentar_Erlaeuterung"))
}

}
