% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_extract_person.R
\name{qb_extract_person}
\alias{qb_extract_person}
\title{Extract name and contact data of a Person.}
\usage{
qb_extract_person(obj, role = NA_character_, names_as_in_database = TRUE)
}
\arguments{
\item{obj}{An XML-element of type "Kontakt_Person_lang" or "Kontakt_Person_kurz".}

\item{role}{A string value which is used for inserting a variable \code{role}
listing the role a person mentioned in the XML file is associated with.
Defaults to \code{NA}.}

\item{names_as_in_database}{A boolean value that if \code{TRUE} indicates that
the result should have column names renamed as they are defined in the database,
which are: \code{idPerson}, \code{role}, \code{title}, \code{firstname},
\code{lastname}, \code{responsibility}, \code{phone}, \code{fax}, and
\code{email}. The variable \code{idPerson} is added as \code{NA_integer}
that represents the primary key in the database table "Person". Default is
\code{TRUE}.}
}
\value{
A tibble with the following columns (all of type "character"): "Titel",
    "Vorname", "Nachname", "Funktion_Arbeitsschwerpunkt", "Telefon", "Fax" and
     "Email". Empty values are encoding as \code{NA_character}.
}
\description{
Extract name and contact data of a Person.
}
\examples{

doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)
x <- xml2::xml_find_all(doc,
        '//Einleitung/Verantwortlicher_Erstellung/Kontakt_Person_lang')

qb_extract_person(x, role = "Verantwortlicher_Erstellung")


}
