% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_extract_selective_lists.R
\name{qb_extract_selective_lists}
\alias{qb_extract_selective_lists}
\title{Extract all lists from an Excel-file with the selective lists provided by the
GBA which are then used in a "Qualitaetsbericht" to encode specific information.}
\usage{
qb_extract_selective_lists(XLS_path = NULL)
}
\arguments{
\item{XLS_path}{The file path to an Excel-file.}
}
\value{
A named list where each element is a tibble reproducing one sheet from
    the original Excel-file. The first tibble contains the table of contents
    of the remaining sheets.
}
\description{
Extract all lists from an Excel-file with the selective lists provided by the
GBA which are then used in a "Qualitaetsbericht" to encode specific information.
}
\examples{

\dontrun{
download.file(
  "https://www.g-ba.de/downloads/40-268-7467/2021-04-07_Qb-R_Anlage-3_Auswahllisten.xlsx",
  destfile = "test.xlsx", method = "libcurl", mode = "wb")

qb_extract_selective_lists("test.xlsx")
}

}
