% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_extract_simple_section.R
\name{qb_extract_simple_section}
\alias{qb_extract_simple_section}
\title{Extract repeated XML-children from a section together with its values.}
\usage{
qb_extract_simple_section(
  obj,
  entrypoint = NA_character_,
  subelements = NA_character_,
  message_missing = FALSE
)
}
\arguments{
\item{obj}{An XML-nodeset that contains several similar children which in turn
contain several variables that one seek to extract. More than this level
of nested structures is not supported.}

\item{entrypoint}{A single string value that specifies the name of the
children elements.}

\item{subelements}{A string vector with the names of the variables that should
be extracted which are then returned as columns in the resulting \code{tibble}.
Here, one can also specify nested elements, i.e., elements that are on a
deeper level than the other subelements. In this way, more complicated
structures with different nesting levels can be extracted into a plain
table structure.}

\item{message_missing}{Boolean value to indicate if a message should be printed
in case a missing obj is found and a dummy tibble with all specified variables
from \code{subelements} as \code{NA_character_}. Defaults to \code{FALSE}, i.e.,
no message is printed.}
}
\value{
A \code{tibble} with as many character columns as the length of the
    parameter \code{subelements}.
}
\description{
Extract repeated XML-children from a section together with its values.
}
\examples{

\dontrun{
doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)


x <- xml2::xml_find_all(doc, '//Einleitung/Weiterfuehrende_Links')
qb_extract_simple_section(x, "Link", c("URL", "Beschreibung"))

x <- xml2::xml_find_all(doc, '//Akademische_Lehre')
qb_extract_simple_section(x, "Akademische_Lehre_Wissenschaftliche_Taetigkeit",
                          c("FL_Schluessel", "Erlaeuterungen"))

x <- xml2::xml_find_all(doc, '//Ausbildung_in_anderen_Heilberufen')
qb_extract_simple_section(x, "Ausbildung_in_anderen_Heilberufen",
                          c("HB_Schluessel", "Erlaeuterungen"))

x <- xml2::xml_find_all(doc, '//Barrierefreiheit')
qb_extract_simple_section(x, "Barrierefreiheit_Aspekt",
                          c("BF_Schluessel", "Erlaeuterungen"))

x <- xml2::xml_find_first(doc, '//Medizinische_Leistungsangebote')
qb_extract_simple_section(x, "Medizinisches_Leistungsangebot",
                          c("VA_VU_Schluessel", "Erlaeuterungen",
                          "Sonstiger/VA_VU_Sonstiger_Schluessel",
                          "Sonstiger/Bezeichnung"))
}

}
