% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_extractor_Allgemeine_Informationen.R
\name{qb_extractor_Allgemeine_Informationen}
\alias{qb_extractor_Allgemeine_Informationen}
\title{Extract all information regarding the topic "Akademische Lehre".}
\usage{
qb_extractor_Allgemeine_Informationen(obj, file_name, Fachabteilungen)
}
\arguments{
\item{obj}{An XML-object comprising a complete detailed report.}

\item{file_name}{A single string with the name of the XML-file that is
currently processed. No default.}

\item{Fachabteilungen}{A \code{data.frame} that comprises information about
all possible departments types. For the exact columns that are needed,
see \link{Fachabteilungen_2019}.}
}
\value{
A list with the following elements:
    \itemize{
        \item HospitalOperator
        \item Krankenhaus_Name
        \item Krankenhaus_IK
        \item Krankenhaus_Standortnummer
        \item Krankenhaus_Adressen
        \item Krankenhaus_Telefon
        \item Krankenhaus_Leitungspersonen
        \item Standort_Name
        \item Standort_IK
        \item Standort_Standortnummer
        \item Standort_Adressen
        \item Standort_Telefon
        \item Standort_Leitungspersonen
        \item Universitaetsklinikum
        \item Akademisches_Lehrkrankenhaus
        \item Name_Universitaet
        \item Psychiatrisches_Krankenhaus
        \item Psychiatrie_Versorgungsverpflichtung
    }
}
\description{
Extract all information regarding the topic "Akademische Lehre".
}
\examples{

\dontrun{
doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)

file_name <- stringr::str_replace(basename(doc_path), "-xml\\\\.xml$", "")

qb_extractor_Allgemeine_Informationen(doc,
                                      file_name = file_name,
                                      Fachabteilungen = Fachabteilungen_2019)
}

}
