% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_extractor_Arzneimitteltherapiesicherheit.R
\name{qb_extractor_Arzneimitteltherapiesicherheit}
\alias{qb_extractor_Arzneimitteltherapiesicherheit}
\title{Extract all information regarding the topic "Arzneimitteltherapiesicherheit".}
\usage{
qb_extractor_Arzneimitteltherapiesicherheit(obj, Hospital_id)
}
\arguments{
\item{obj}{An XML-object comprising a complete detailed report.}

\item{Hospital_id}{An integer value that represents the primary key of a
hospital in the designated database.}
}
\value{
A list with the following elements:
    \itemize{
        \item Arzneimittelkommission
        \item PharmacotherapySteeringBoard_exists
        \item PharmacotherapySteeringBoard
        \item Kein_Gremium_Arbeitsgruppe
        \item Verantwortliche_Person_AMTS
        \item Verantwortliche_Person_AMTS_nicht_festgelegt
        \item pharmaceuticalStaff_exists
        \item pharmaceuticalStaff
        \item PharmacotherapySafetyMeasures_exists
        \item PharmacotherapySafetyMeasures
        \item InstrumentsRM_exists
        \item InstrumentsRM
    }
}
\description{
Extract all information regarding the topic "Arzneimitteltherapiesicherheit".
}
\examples{

doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)

qb_extractor_Arzneimitteltherapiesicherheit(doc, 1L)


}
