% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_extractor_Einleitung.R
\name{qb_extractor_Einleitung}
\alias{qb_extractor_Einleitung}
\title{Extract all information regarding the topic "Einleitung".}
\usage{
qb_extractor_Einleitung(obj)
}
\arguments{
\item{obj}{An XML-object comprising a complete detailed report.}
}
\value{
A list with the following elements:
    \itemize{
        \item Dataset_metadata
        \item URL_Internetseite_Krankenhaus
        \item URL_Weitere_Informationen
        \item Further_links
        \item Verantwortlicher_Erstellung_Qualitaetsbericht
        \item Verantwortlicher_Krankenhausleitung
    }
}
\description{
Extract all information regarding the topic "Einleitung".
}
\examples{

doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)

qb_extractor_Einleitung(doc)


}
