% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_extractor_Mindestmengen.R
\name{qb_extractor_Mindestmengen}
\alias{qb_extractor_Mindestmengen}
\title{Extract all information regarding the topic "Mindestmengen".}
\usage{
qb_extractor_Mindestmengen(obj, Hospital_id)
}
\arguments{
\item{obj}{An XML-object comprising a complete detailed report.}

\item{Hospital_id}{An integer value that represents the primary key of a
hospital in the designated database.}
}
\value{
A list with the following elements:
    \itemize{
        \item MinimumQuantities_exists
        \item MinimumQuantities
        \item MinimumQuantitiesPrognosis_exists
        \item MinimumQuantitiesPrognosis
    }
}
\description{
Extract all information regarding the topic "Mindestmengen".
}
\examples{

doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)

qb_extractor_Mindestmengen(doc, 1L)


}
