% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/qb_extractor_Nicht_Medizinische_Leistungsangebote.R
\name{qb_extractor_Nicht_Medizinische_Leistungsangebote}
\alias{qb_extractor_Nicht_Medizinische_Leistungsangebote}
\title{Extract all information regarding the topic "Nicht Medizinische Leistungsangebote".}
\usage{
qb_extractor_Nicht_Medizinische_Leistungsangebote(obj, Hospital_id)
}
\arguments{
\item{obj}{An XML-object comprising a complete detailed report.}

\item{Hospital_id}{An integer value that represents the primary key of a
hospital in the designated database.}
}
\value{
A list with the following elements:
    \itemize{
        \item NonMedicalServices_exists
        \item NonMedicalServices
    }
}
\description{
Extract all information regarding the topic "Nicht Medizinische Leistungsangebote".
}
\examples{

doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)

qb_extractor_Nicht_Medizinische_Leistungsangebote(doc, 1L)


}
