% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_extractor_OE.R
\name{qb_extractor_OE}
\alias{qb_extractor_OE}
\title{Extract all information from all departments in one hospital.}
\usage{
qb_extractor_OE(obj, year)
}
\arguments{
\item{obj}{An XML-object comprising not more than one department within a
hospital.}

\item{year}{A character value with the name of the year of the report, which
is needed because of changes of XML-elements' names between years.}
}
\value{
A list with the following elements:
    \itemize{
        \item OE_Stammdaten
        \item Aerztliche_Leitung_OE
        \item Kontakt_OE_Adresse
        \item FA_Schluessel
        \item Medizinische_Leistungsangebote
        \item Hauptdiagnosen
        \item Prozeduren
        \item Ambulante_Behandlungsmoeglichkeiten
        \item Ambulanz_116b
        \item Ambulante_Operationen
        \item Personal_OE
        \item Aerztliche_Fachexpertisen
        \item Pflegerische_Fachexpertisen
    }
}
\description{
Extract all information from all departments in one hospital.
}
\examples{

\dontrun{
doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)

objOEs <- xml_find_all(doc,
    '//Organisationseinheiten_Fachabteilungen/Organisationseinheit_Fachabteilung')

qb_extractor_OE(objOEs[[1]])

all_OEs <- map(objOEs, ~ qb_extractor_OE(.x))
}

}
