% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_extractor_Personal_des_Krankenhauses.R
\name{qb_extractor_Personal_des_Krankenhauses}
\alias{qb_extractor_Personal_des_Krankenhauses}
\title{Extract all information regarding the topic "Personal des Krankenhauses".}
\usage{
qb_extractor_Personal_des_Krankenhauses(obj, Hospital_id)
}
\arguments{
\item{obj}{An XML-object comprising a complete detailed report.}

\item{Hospital_id}{An integer value that represents the primary key of a
hospital in the designated database.}
}
\value{
A list with the following elements:
    \itemize{
        \item PersonnelHospital_exists
        \item PersonnelHospital
        \item AttendingDoctorsHospital_exists
        \item AttendingDoctorsHospital
        \item SpecialPersonnelHospital_exists
        \item SpecialPersonnelHospital
        \item tariffWeeklyWorkingHoursDoctors
        \item tariffWeeklyWorkingHoursNurses
    }
}
\description{
Extract all information regarding the topic "Personal des Krankenhauses".
}
\examples{

doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)

qb_extractor_Personal_des_Krankenhauses(doc, 1L)


}
