% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/qb_extractor_Umgang_mit_Risiken_in_der_Patientenversorgung.R
\name{qb_extractor_Umgang_mit_Risiken_in_der_Patientenversorgung}
\alias{qb_extractor_Umgang_mit_Risiken_in_der_Patientenversorgung}
\title{Extract all information regarding the topic "Umgang mit Risiken in der Patientenversorgung".}
\usage{
qb_extractor_Umgang_mit_Risiken_in_der_Patientenversorgung(obj, Hospital_id)
}
\arguments{
\item{obj}{An XML-object comprising a complete detailed report.}

\item{Hospital_id}{An integer value that represents the primary key of a
hospital in the designated database.}
}
\value{
A list with the following elements:
    \itemize{
        \item Qualitymanagement_exists
        \item Qualitymanagement
        \item Riskmanagement_exists
        \item Riskmanagement
        \item Persons_QM_RM
    }
}
\description{
Extract all information regarding the topic "Umgang mit Risiken in der Patientenversorgung".
}
\examples{

doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)

qb_extractor_Umgang_mit_Risiken_in_der_Patientenversorgung(doc, 1L)


}
