% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_extractor_Weitere_Informationen_Hygiene.R
\name{qb_extractor_Weitere_Informationen_Hygiene}
\alias{qb_extractor_Weitere_Informationen_Hygiene}
\title{Extract all information regarding the topic "Weitere Informationen Hygiene".}
\usage{
qb_extractor_Weitere_Informationen_Hygiene(obj, Hospital_id, year = NULL)
}
\arguments{
\item{obj}{An XML-object comprising a complete detailed report.}

\item{Hospital_id}{An integer value that represents the primary key of a
hospital in the designated database.}

\item{year}{A character value with the name of the year of the report, which
is needed because of changes of XML-elements' names between years.}
}
\value{
A list with the following elements:
    \itemize{
        \item HygieneFurtherMeasures_exists
        \item HygieneFurtherMeasures
        \item HygieneMeasures_exists
        \item HygieneMeasures
    }
}
\description{
Extract all information regarding the topic "Weitere Informationen Hygiene".
}
\examples{

doc_path <- system.file("extdata", "260100023-01-2019-xml.xml",
                        package = "qbgbaReader", mustWork = TRUE)
doc <- xml2::read_xml(doc_path)

qb_extractor_Weitere_Informationen_Hygiene(doc, 1L, "2019")


}
